Travis Savo's patches:
1.  Made listener id long, to prevent duplication after lots of reconnects.

2.  Fixed potential looping bug in remote cache removeAll


git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@224089 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/jcs/auxiliary/disk/AbstractDiskCache.java
@@ -354,12 +354,12 @@ public abstract class AbstractDiskCache implements AuxiliaryCache, Serializable
      */
     private class MyCacheListener implements ICacheListener
     {
-        private byte listenerId = 0;
+        private long listenerId = 0;
 
         /**
          * @see ICacheListener#getListenerId
          */
-        public byte getListenerId()
+        public long getListenerId()
             throws IOException
         {
             return this.listenerId;
@@ -368,7 +368,7 @@ public abstract class AbstractDiskCache implements AuxiliaryCache, Serializable
         /**
          * @see ICacheListener#setListenerId
          */
-        public void setListenerId( byte id )
+        public void setListenerId( long id )
             throws IOException
         {
             this.listenerId = id;
