+++ src/java/org/apache/jcs/auxiliary/lateral/LateralCacheInfo.java
@@ -40,6 +40,6 @@ public class LateralCacheInfo
      */
     protected static VMID vmid = new VMID();
     /** Description of the Field */
-    public static byte listenerId = ( byte ) vmid.hashCode();
+    public static long listenerId = vmid.hashCode();
 
 }
