+++ src/java/org/apache/jcs/auxiliary/lateral/javagroups/LateralJGService.java
@@ -101,7 +101,7 @@ public class LateralJGService
      * @param requesterId
      * @exception IOException
      */
-    public void update( ICacheElement item, byte requesterId )
+    public void update( ICacheElement item, long requesterId )
         throws IOException
     {
         LateralElementDescriptor led = new LateralElementDescriptor( item );
@@ -127,7 +127,7 @@ public class LateralJGService
      * @param requesterId
      * @exception IOException
      */
-    public void remove( String cacheName, Serializable key, byte requesterId )
+    public void remove( String cacheName, Serializable key, long requesterId )
         throws IOException
     {
         CacheElement ce = new CacheElement( cacheName, key, null );
@@ -192,7 +192,7 @@ public class LateralJGService
      * @param requesterId
      * @exception IOException
      */
-    public void removeAll( String cacheName, byte requesterId )
+    public void removeAll( String cacheName, long requesterId )
         throws IOException
     {
         CacheElement ce = new CacheElement( cacheName, "ALL", null );
