+++ src/java/org/apache/jcs/auxiliary/lateral/socket/tcp/LateralTCPSender.java
@@ -312,7 +312,7 @@ public class LateralTCPSender
 
     // Service Methods //
     /** Description of the Method */
-    public void update( ICacheElement item, byte requesterId )
+    public void update( ICacheElement item, long requesterId )
         throws IOException
     {
         LateralElementDescriptor led = new LateralElementDescriptor( item );
@@ -331,7 +331,7 @@ public class LateralTCPSender
 
 
     /** Description of the Method */
-    public void remove( String cacheName, Serializable key, byte requesterId )
+    public void remove( String cacheName, Serializable key, long requesterId )
         throws IOException
     {
         CacheElement ce = new CacheElement( cacheName, key, null );
@@ -373,7 +373,7 @@ public class LateralTCPSender
 
 
     /** Description of the Method */
-    public void removeAll( String cacheName, byte requesterId )
+    public void removeAll( String cacheName, long requesterId )
         throws IOException
     {
         CacheElement ce = new CacheElement( cacheName, "ALL", null );
