+++ src/java/org/apache/jcs/auxiliary/lateral/socket/tcp/LateralTCPService.java
@@ -94,7 +94,7 @@ public class LateralTCPService
     }
 
     /** */
-    public void update( ICacheElement item, byte requesterId )
+    public void update( ICacheElement item, long requesterId )
         throws IOException
     {
         LateralElementDescriptor led = new LateralElementDescriptor( item );
@@ -111,7 +111,7 @@ public class LateralTCPService
     }
 
     /** */
-    public void remove( String cacheName, Serializable key, byte requesterId )
+    public void remove( String cacheName, Serializable key, long requesterId )
         throws IOException
     {
         CacheElement ce = new CacheElement( cacheName, key, null );
@@ -164,10 +164,10 @@ public class LateralTCPService
      */
     public Set getGroupKeys(String cacheName, String group)
     {
-        if (true) 
+        if (true)
         {
             throw new UnsupportedOperationException("Groups not implemented.");
-        }        
+        }
         return null;
     }
 
@@ -179,7 +179,7 @@ public class LateralTCPService
     }
 
     /** */
-    public void removeAll( String cacheName, byte requesterId )
+    public void removeAll( String cacheName, long requesterId )
         throws IOException
     {
         CacheElement ce = new CacheElement( cacheName, "ALL", null );
