+++ src/java/org/apache/jcs/auxiliary/remote/ZombieRemoteCacheService.java
@@ -34,15 +34,15 @@ public class ZombieRemoteCacheService extends ZombieCacheService implements IRem
 {
 
     /** Description of the Method */
-    public void update( ICacheElement item, byte listenerId ) { }
+    public void update( ICacheElement item, long listenerId ) { }
 
 
     /** Description of the Method */
-    public void remove( String cacheName, Serializable key, byte listenerId ) { }
+    public void remove( String cacheName, Serializable key, long listenerId ) { }
 
 
     /** Description of the Method */
-    public void removeAll( String cacheName, byte listenerId ) { }
+    public void removeAll( String cacheName, long listenerId ) { }
 
     public Set getGroupKeys(String cacheName, String groupName)
     {
