+++ src/java/org/apache/jcs/auxiliary/remote/behavior/IRemoteCacheService.java
@@ -37,17 +37,17 @@ public interface IRemoteCacheService extends Remote, ICacheService
 {
 
     /** Puts a cache item to the cache. */
-    public void update( ICacheElement item, byte requesterId )
+    public void update( ICacheElement item, long requesterId )
         throws ObjectExistsException, IOException;
 
 
     /** Removes the given key from the specified cache. */
-    public void remove( String cacheName, Serializable key, byte requesterId )
+    public void remove( String cacheName, Serializable key, long requesterId )
         throws IOException;
 
 
     /** Remove all keys from the sepcified cache. */
-    public void removeAll( String cacheName, byte requesterId )
+    public void removeAll( String cacheName, long requesterId )
         throws IOException;
 
     public Set getGroupKeys(String cacheName, String groupName)
