+++ src/java/org/apache/jcs/auxiliary/remote/server/RemoteCacheServerInfo.java
@@ -40,6 +40,6 @@ public class RemoteCacheServerInfo
      */
     protected static VMID vmid = new VMID();
     /** Description of the Field */
-    public static byte listenerId = ( byte ) vmid.hashCode();
+    public static long listenerId = ( long ) vmid.hashCode();
 
 }
