+++ src/java/org/apache/jcs/engine/behavior/ICacheListener.java
@@ -63,7 +63,7 @@ public interface ICacheListener
      *
      * @param id The new listenerId value
      */
-    public void setListenerId( byte id )
+    public void setListenerId( long id )
         throws IOException;
 
 
@@ -72,7 +72,7 @@ public interface ICacheListener
      *
      * @return The listenerId value
      */
-    public byte getListenerId()
+    public long getListenerId()
         throws IOException;
 
 }
